✨ fonctionnalité(DangerZone.js) : créer un composant DangerZone pour gérer les actions dangereuses 🔀 refactor(Nav.js) : supprimer l'import et l'utilisation du composant DangerZone de l'emplacement précédent 🔀 refactor(DangerZone.js) : déplacer le composant DangerZone vers le dossier Global 🔀 refactor(profil.js) : mettre à jour l'import du composant DangerZone
